楽譜データの名前をScoreにして型をtype.tsに移動
